+++ src/PatronBlockModal.js
@@ -1,4 +1,4 @@
-import { take, orderBy } from 'lodash';
+import { take, orderBy, isObject } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
@@ -10,21 +10,14 @@ import {
 } from '@folio/stripes/components';
 
 const PatronBlockModal = ({ open, onClose, patronBlocks, automatedPatronBlocks, viewUserPath }) => {
-  const blocks = take(orderBy(patronBlocks, ['metadata.updatedDate'], ['desc']), 1);
-  const renderBlocks = blocks.map(block => {
+  const orderedPatronBlocks = orderBy(patronBlocks, ['metadata.updatedDate'], ['desc']);
+  const blocks = [...automatedPatronBlocks, ...orderedPatronBlocks];
+
+  const blocksToRender = take(blocks, 3).map(block => {
     return (
       <Row>
         <Col xs>
-          <b>{block.desc || ''}</b>
-        </Col>
-      </Row>
-    );
-  });
-  const renderAutomatedPatronBlocks = automatedPatronBlocks.map(block => {
-    return (
-      <Row key={block}>
-        <Col xs>
-          <b>{block}</b>
+          <b>{ isObject(block) ? (block.desc || '') : block }</b>
         </Col>
       </Row>
     );
@@ -49,11 +42,10 @@ const PatronBlockModal = ({ open, onClose, patronBlocks, automatedPatronBlocks, 
           :
         </Col>
       </Row>
-      {renderAutomatedPatronBlocks}
-      {renderBlocks}
+      {blocksToRender}
       <br />
       <Row>
-        <Col xs={8}>{(patronBlocks.length > 3) && <FormattedMessage id="ui-requests.additionalReasons" />}</Col>
+        <Col xs={8}>{(blocks.length > 3) && <FormattedMessage id="ui-requests.additionalReasons" />}</Col>
         <Col xs={4}>
           <Row end="xs">
             <Col>